requirements.txt
dicttoxml==1.3.1
Django==1.11.2
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-geojson==2.10.0
django-import-export==0.4.3
django-leaflet==0.22.0
django-localflavor==1.5.1
django-mathfilters==0.4.0
django-multiupload==0.5.2
django-registration==2.2
django-report-utils==0.3.1
django-admin-report==0.4
django-rest-swagger==2.1.2
django-tables2==1.9.0
djangorestframework==3.6.3
django-filter==1.0.4
httplib2==0.10.3
jdcal==1.3
macholib==1.8
openpyxl==2.4.8
Pillow==2.5.1
python-dateutil==2.6.0
requests==2.18.1
django-ckeditor==5.2.2
django-wysiwyg==0.8.0
MySQL-python==1.2.5
django-uuidfield==0.5.0
simplejson==3.11.1
python-logstash==0.4.6
django-simple-history==1.9.0
requests-oauthlib==0.8.0
pytz==2017.2
social-auth-app-django==1.2.0
PyYAML==3.12
reportlab==3.4.0
WeasyPrint==0.39
ldap==1.0.2
python-ldap==2.4.44
